avoid lock contention
