+++ src/stream/src/executor/managed_state/join/mod.rs
@@ -173,7 +173,7 @@ impl<K: HashKey, S: StateStore> JoinHashMap<K, S> {
     /// Returns a mutable reference to the value of the key in the memory, if does not exist, look
     /// up in remote storage and return, if still not exist, return None.
     #[allow(dead_code)]
-    pub async fn get<'a>(&'a mut self, key: &K) -> Option<&'a HashValueType<S>> {
+    pub async fn get<'a, 'b>(&'a mut self, key: &'b K) -> Option<&'a HashValueType<S>> {
         let state = self.inner.get(key);
         // TODO: we should probably implement a entry function for `LruCache`
         match state {
@@ -208,7 +208,7 @@ impl<K: HashKey, S: StateStore> JoinHashMap<K, S> {
     /// Returns a mutable reference to the value of the key in the memory, if does not exist, look
     /// up in remote storage and return the [`JoinEntryState`] without cached state, if still not
     /// exist, return None.
-    pub async fn get_mut_without_cached<'a, 'b>(
+    pub async fn get_mut_without_cached<'a, 'b: 'a>(
         &'a mut self,
         key: &'b K,
     ) -> RwResult<Option<&'a mut HashValueType<S>>> {
