+++ src/stream/src/executor/receiver.rs
@@ -16,7 +16,7 @@ use futures::channel::mpsc::Receiver;
 use futures::StreamExt;
 use risingwave_common::catalog::Schema;
 
-use super::{ActorContextRef, OperatorInfo};
+use super::{ActorContextRef, OperatorInfoStatus};
 use crate::executor::{
     BoxedMessageStream, Executor, ExecutorInfo, Message, PkIndices, PkIndicesRef,
 };
@@ -26,11 +26,12 @@ use crate::executor::{
 /// messages down to the executors.
 pub struct ReceiverExecutor {
     receiver: Receiver<Message>,
+
     /// Logical Operator Info
     info: ExecutorInfo,
-    /// Actor context
-    actor_context: ActorContextRef,
-    actor_context_position: usize,
+
+    /// Actor operator context
+    status: OperatorInfoStatus,
 }
 
 impl std::fmt::Debug for ReceiverExecutor {
@@ -50,13 +51,6 @@ impl ReceiverExecutor {
         actor_context: ActorContextRef,
         receiver_id: u64,
     ) -> Self {
-        let actor_context_position = {
-            let mut ctx = actor_context.lock();
-            let actor_context_position = ctx.info.len();
-            ctx.info.push(OperatorInfo::new(receiver_id));
-            actor_context_position
-        };
-
         Self {
             receiver,
             info: ExecutorInfo {
@@ -64,19 +58,17 @@ impl ReceiverExecutor {
                 pk_indices,
                 identity: "ReceiverExecutor".to_string(),
             },
-            actor_context,
-            actor_context_position,
+            status: OperatorInfoStatus::new(actor_context, receiver_id),
         }
     }
 }
 
 impl Executor for ReceiverExecutor {
     fn execute(self: Box<Self>) -> BoxedMessageStream {
-        let actor_context = self.actor_context.clone();
-        let actor_context_position = self.actor_context_position;
+        let mut status = self.status;
         self.receiver
             .map(move |msg| {
-                actor_context.lock().info[actor_context_position].next_message(&msg);
+                status.next_message(&msg);
                 Ok(msg)
             })
             .boxed()
